tax toggles, isolate summary component
