+++ src/lib/utils/utility.ts
@@ -1,0 +1,8 @@
+export function formatCurrency(value: number) {
+    return new Intl.NumberFormat('en-US', {
+      style: 'currency',
+      currency: 'USD',
+      minimumFractionDigits: 2,
+      maximumFractionDigits: 2
+    }).format(value);
+  }
